requirements.txt
streamlit
google-genai
youtube-transcript-api
textblob
pandas